Added msgNumber
The token now has a message number which is created when the token is
created. This is sent in the token to the user. Currently the UI then
sends that msgnumber to the two clients, but im going to change that so
they just pull it out of the token.